Conflicts:
	Clap/src/main/java/product/model/ProductVO.java
	Clap/src/main/webapp/sql/SQLQuery2.sql

+++ Clap/src/main/java/shopping/model/OrderFormVO.java
@@ -30,7 +30,7 @@ public class OrderFormVO implements Serializable  {
     @Column(name="orderform_time")
 	private Timestamp time;
     @Column(name="orderform_total")
-	private Integer total;
+	private Double total;
     @Column(name="orderform_status")
 	private Byte status;
     @ManyToOne
@@ -59,10 +59,10 @@ public class OrderFormVO implements Serializable  {
 	public void setTime(Timestamp time) {
 		this.time = time;
 	}
-	public Integer getTotal() {
+	public Double getTotal() {
 		return total;
 	}
-	public void setTotal(Integer total) {
+	public void setTotal(Double total) {
 		this.total = total;
 	}
 	public Byte getStatus() {
